requirements.txt
#
#Módulo dotenv para las variables .env
#
python-dotenv==0.15.0
#Alternativa para el módulo  nv
#python3 -m pip install nv
#
#Módulo requests
requests==2.26.0
#Módulo para la autentificación de Spotify para métodos ya creados 
spotipy==2.18.0
